chartsapi - Makes modular
