+++ amplitude_python_sdk/v2/__init__.py
@@ -2,8 +2,14 @@
 Contains all clients and models pertaining to the Amplitude V2 API.
 """
 
-from .client import AmplitudeV2APIClient
+from amplitude_python_sdk.v2.clients.http_batch_client import AmplitudeV2APIClient
+from amplitude_python_sdk.v2.clients.chart_annotations_client import ChartAnnotationsAPIClient
 from .models.event import EventIdentifiers, EventLocationData, EventV2
 from .models.charts import ChartAnnotationsV2
 
-__all__ = ["AmplitudeV2APIClient", "EventIdentifiers", "EventLocationData", "EventV2", "ChartAnnotationsV2"]
+__all__ = ["AmplitudeV2APIClient",
+           "ChartAnnotationsAPIClient",
+           "EventIdentifiers",
+           "EventLocationData",
+           "EventV2",
+           "ChartAnnotationsV2"]
