+++ amplitude_python_sdk/v2/clients/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains all client methods for connecting to Amplitude V2 API's
+"""
